Remake the ConfigurationDialog to the ImportDefDialog
