--- src/components/configure/configuration/ConfigurationDialog.tsx
@@ -1,171 +1,0 @@
-import React from 'react';
-import Draggable from 'react-draggable';
-import './ConfigurationDialog.scss';
-import {
-  ConfigurationDialogActionsType,
-  ConfigurationDialogStateType,
-} from './ConfigurationDialog.container';
-import { Button, Dialog, DialogContent, DialogTitle } from '@material-ui/core';
-import CloseIcon from '@material-ui/icons/Close';
-import Paper, { PaperProps } from '@material-ui/core/Paper';
-import Tabs from '@material-ui/core/Tabs';
-import Tab from '@material-ui/core/Tab';
-import { Alert, AlertTitle } from '@material-ui/lab';
-
-import { KeyboardDefinitionSchema } from '../../../gen/types/KeyboardDefinition';
-import { KeyboardDefinitionFormPart } from '../../common/keyboarddefformpart/KeyboardDefinitionFormPart';
-
-type OwnProps = {
-  open: boolean;
-  onClose: () => void;
-  vendorId: number;
-  productId: number;
-  productName: string;
-};
-
-type ConfigurationDialogProps = OwnProps &
-  Partial<ConfigurationDialogActionsType> &
-  Partial<ConfigurationDialogStateType>;
-
-type OwnState = {
-  selectedMenuIndex: number;
-  keyboardDefinition: KeyboardDefinitionSchema | null;
-  keyboardDefinitionFile: string | null;
-};
-
-export default class ConfigurationDialog extends React.Component<
-  ConfigurationDialogProps,
-  OwnState
-> {
-  constructor(
-    props: ConfigurationDialogProps | Readonly<ConfigurationDialogProps>
-  ) {
-    super(props);
-    this.state = {
-      selectedMenuIndex: 0,
-      keyboardDefinition: null,
-      keyboardDefinitionFile: null,
-    };
-  }
-
-  private clearKeyboardDefinition() {
-    this.setState({
-      keyboardDefinition: null,
-      keyboardDefinitionFile: null,
-    });
-  }
-
-  private onClickApplyKeyboardDefinition() {
-    this.props.refreshKeyboardDefinition!(this.state.keyboardDefinition!);
-    this.clearKeyboardDefinition();
-  }
-
-  // eslint-disable-next-line no-unused-vars
-  private onLoadFile(
-    keyboardDefinition: KeyboardDefinitionSchema,
-    name: string
-  ) {
-    this.setState({
-      keyboardDefinition: keyboardDefinition,
-      keyboardDefinitionFile: name,
-    });
-  }
-
-  render() {
-    let panelIndex = 0;
-    return (
-      <Dialog
-        open={this.props.open}
-        maxWidth={'md'}
-        PaperComponent={PaperComponent}
-        className="layout-options-dialog"
-      >
-        <DialogTitle id="draggable-dialog-title" style={{ cursor: 'move' }}>
-          Configuration
-          <div className="close-dialog">
-            <CloseIcon onClick={this.props.onClose} />
-          </div>
-        </DialogTitle>
-        <DialogContent dividers className="layout-options">
-          <Tabs
-            orientation="vertical"
-            variant="scrollable"
-            value={this.state.selectedMenuIndex}
-            onChange={(_, newValue) => {
-              this.setState({ selectedMenuIndex: newValue });
-            }}
-            className="config-menu"
-          >
-            <Tab label="Import" />
-          </Tabs>
-          <TabPanel value={this.state.selectedMenuIndex} index={panelIndex++}>
-            <KeyboardDefinitionFormPart
-              messageHtml={`Please import <strong>${this.props.productName}</strong>'s defintion file (.json).`}
-              validateDeviceIds={true}
-              deviceVendorId={this.props.vendorId}
-              deviceProductId={this.props.productId}
-              onLoadFile={(kd, name) => {
-                this.onLoadFile(kd, name);
-              }}
-              size="small"
-            />
-
-            {this.state.keyboardDefinition && (
-              <Alert severity="success" className="import-success">
-                <AlertTitle>{`Valid (${this.state.keyboardDefinitionFile})`}</AlertTitle>
-                {`Will you apply the keyboard definition file?`}
-                <div className="apply-definition">
-                  <Button
-                    size="small"
-                    variant="text"
-                    color="primary"
-                    disableElevation
-                    onClick={this.clearKeyboardDefinition.bind(this)}
-                  >
-                    Cancel
-                  </Button>
-                  <Button
-                    size="small"
-                    variant="contained"
-                    color="primary"
-                    disableElevation
-                    onClick={this.onClickApplyKeyboardDefinition.bind(this)}
-                  >
-                    Apply
-                  </Button>
-                </div>
-              </Alert>
-            )}
-          </TabPanel>
-        </DialogContent>
-      </Dialog>
-    );
-  }
-}
-
-function TabPanel(props: { value: number; index: number; children: any }) {
-  const { children, value, index } = props;
-
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`vertical-tabpanel-${index}`}
-      aria-labelledby={`vertical-tab-${index}`}
-      className="tab-panel"
-    >
-      {value === index && <React.Fragment>{children}</React.Fragment>}
-    </div>
-  );
-}
-
-function PaperComponent(props: PaperProps) {
-  return (
-    <Draggable
-      handle="#draggable-dialog-title"
-      cancel={'[class*="MuiDialogContent-root"]'}
-    >
-      <Paper {...props} />
-    </Draggable>
-  );
-}
