+++ src/components/configure/importDef/ImportDefDialog.stories.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import CssBaseline from '@material-ui/core/CssBaseline';
-import ConfigurationDialog from './ConfigurationDialog';
+import ImportDefDialog from './ImportDefDialog';
 import { KeyboardDefinitionSchema } from '../../../gen/types/KeyboardDefinition';
 import { hexadecimal } from '../../../utils/StringUtils';
 
 export default {
-  title: 'ConfigurationDialog',
-  component: ConfigurationDialog,
+  title: 'ImportDefDialog',
+  component: ImportDefDialog,
   decorators: [
     (Story: any) => (
       <React.Fragment>
@@ -62,7 +62,7 @@ const keyboardDefinition: KeyboardDefinitionSchema = {
 };
 
 export const Default = () => (
-  <ConfigurationDialog
+  <ImportDefDialog
     open={true}
     onClose={() => {}}
     vendorId={0x5954}
@@ -73,7 +73,7 @@ export const Default = () => (
 );
 
 export const NoOptions = () => (
-  <ConfigurationDialog
+  <ImportDefDialog
     open={true}
     onClose={() => {}}
     vendorId={0x5954}
@@ -85,13 +85,12 @@ export const NoOptions = () => (
   />
 );
 
-class ConfigurationDialogAdapter extends React.Component<any, {}> {
-  private dialog: ConfigurationDialog;
+class ImportDefDialogAdapter extends React.Component<any, {}> {
+  private dialog: ImportDefDialog;
   constructor(props: any) {
     super(props);
-    this.dialog = new ConfigurationDialog(props);
+    this.dialog = new ImportDefDialog(props);
     this.dialog.state = {
-      selectedMenuIndex: 1,
       keyboardDefinition: {} as KeyboardDefinitionSchema,
       keyboardDefinitionFile: 'keyboard_def.json',
     };
@@ -102,7 +101,7 @@ class ConfigurationDialogAdapter extends React.Component<any, {}> {
 }
 
 export const ValidKeyboardDefinition = () => (
-  <ConfigurationDialogAdapter
+  <ImportDefDialogAdapter
     open={true}
     onClose={() => {}}
     vendorId={0x5954}
